Use HSET for confirm otp

+++ backend/registration_api/config/config.go
@@ -9,7 +9,7 @@ var Config = struct {
 		PublicKey            string `yaml:"publickey" env:"AUTH_PUBLIC_KEY"`
 		PrivateKey           string `yaml:"privatekey" env:"AUTH_PRIVATE_KEY"`
 		TTLForOtp            int    `yaml:"ttlforotpinminutes"`
-		MAXOtpVerifyAttempts int    `yaml:"maxotpverifyattempts"`
+		MAXOtpVerifyAttempts int64  `yaml:"maxotpverifyattempts"`
 		OTPLength            int    `yaml:"otp_length" env:"OTP_LENGTH" default:"6"`
 	}
